requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==1.10.15
pydantic[email]==1.10.15
google-auth==2.35.0
python-multipart==0.0.9
pip install fastapi uvicorn[standard] sqlalchemy passlib[bcrypt] python-jose[cryptography] email-validator python-multipart
